Cast values.

+++ kronolith/ajax.php
@@ -9,6 +9,7 @@
  *
  * @author  Michael Slusarz <slusarz@horde.org>
  * @author  Jan Schneider <jan@horde.org>
+ * @author  Gonçalo Queirós <mail@goncaloqueiros.net>
  * @package Kronolith
  */
 
@@ -130,7 +131,7 @@ try {
         }
 
         $taskList = Horde_Util::getFormData('list');
-        $taskType = Horde_Util::getFormData('taskType');
+        $taskType = (int)Horde_Util::getFormData('taskType');
         $tasks = $registry->call('tasks/listTasks',
                                  array(null, null, null, $taskList, $taskType, true));
         if (is_a($tasks, 'PEAR_Error')) {
@@ -338,7 +339,7 @@ try {
             break;
         }
         $taskList = Horde_Util::getFormData('taskList');
-        $taskType = Horde_Util::getFormData('taskType');
+        $taskType = (int)Horde_Util::getFormData('taskType');
         $taskId = Horde_Util::getFormData('taskId');
         $saved = $registry->call('tasks/toggleCompletion',
                                  array($taskId, $taskList));
